% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_trend.R
\name{est_trend}
\alias{est_trend}
\title{Estimate Trend}
\usage{
est_trend(input_vec, year_vec = 1960:2020, window_val = 10)
}
\arguments{
\item{input_vec}{A numeric vector of life expectancies across different years.}

\item{year_vec}{A numeric vector of the years observed. Default 1960:2020,}

\item{window_val}{A numeric giving length of window. Default 10.}
}
\value{
A numeric vector.
}
\description{
This function estimates trend in life expectancy over a moving window.
}
